intel: Make brw_ubo_range::block 8 bits
When we originally built the UBO range analysis pass we supposed that we
would want to use it in a Vulkan world where block might be fairly
large.  However, in reality, we only ever use it on things that got
assigned real binding table entries so it's guaranteed to fit in 8 bits.

+++ src/intel/compiler/brw_compiler.h
@@ -570,7 +570,7 @@ struct brw_image_param {
 
 struct brw_ubo_range
 {
-   uint16_t block;
+   uint8_t block;
    uint8_t start;
    uint8_t length;
 };
